+++ example/flutter_example/lib/blocs/slider/slider.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 /// This bloc handles the state of a material [Slider] widget. In particular,
-/// [SliderBloc] stores the currently selected value of the slider
-class SliderBloc extends Bloc<double, double> {
+/// [SliderCubit] stores the currently selected value of the slider.
+class SliderCubit extends Cubit<double> {
   /// The minimum value of the slider.
   final double minValue;
 
@@ -13,17 +13,17 @@ class SliderBloc extends Bloc<double, double> {
   /// The default value of the slider.
   final double current;
 
-  /// By default, the minimum value of the slider is `2`
-  SliderBloc({
+  /// Creates a [Cubit] that keeps track of the state of a slider widget.
+  SliderCubit({
     required this.minValue,
     required this.maxValue,
     required this.current,
   }) : super(current);
 
-  @override
-  Stream<double> mapEventToState(double event) async* {
-    if ((event >= minValue) && (event <= maxValue)) {
-      yield event;
+  /// Updates the current slider position.
+  void updateSlider(double newValue) {
+    if ((newValue >= minValue) && (newValue <= maxValue)) {
+      emit(newValue);
     }
   }
 }