+++ test/utils/expression_parser_test.dart
@@ -21,12 +21,31 @@ void main() {
         parser.evaluate('e'),
         const MoreOrLessEquals(2.7182818284, precision: 1.0e-10),
       );
+      expect(
+        parser.evaluate('G'),
+        const MoreOrLessEquals(0.8346268416, precision: 1.0e-10),
+      );
+    });
+
+    test(
+        "Making sure that 'ExpressionParser' throws when 'evaluate' has the "
+        "'x' variable in the string.", () {
+      const parser = ExpressionParser();
+
+      expect(
+        () => parser.evaluate('5*x*3-4'),
+        throwsA(isA<ExpressionParserException>()),
+      );
+      expect(
+        () => parser.evaluate('x-3'),
+        throwsA(isA<ExpressionParserException>()),
+      );
     });
 
     test("Making sure that 'ExpressionParser' works with functions.", () {
       const parser = ExpressionParser();
 
-      expect(parser.evaluate('sqrt(49)'), equals(7));
+      expect(parser.evaluate('sqrt(49)'), equals(7),);
       expect(
         parser.evaluate('sin(pi)'),
         const MoreOrLessEquals(0, precision: 1.0e-1),
@@ -38,10 +57,6 @@ void main() {
       expect(
         parser.evaluate('tan(pi/4)'),
         const MoreOrLessEquals(1, precision: 1.0e-1),
-      );
-      expect(
-        parser.evaluate('exp(log(5))'),
-        const MoreOrLessEquals(5, precision: 1.0e-1),
       );
       expect(
         parser.evaluate('asin(1)'),
@@ -69,6 +84,24 @@ void main() {
         "Making sure that 'ExpressionParser' works with the 'x' variable "
         "even if 'x' is not present.", () {
       expect(const ExpressionParser().evaluateOn('6*3 + 4', 0), equals(22));
+    });
+
+    test("Making sure that the extension method on 'String' work correctly.",
+        () {
+      expect('6*3 + 4'.isRealFunction, isTrue);
+      expect('6*3 + 4'.isNumericalExpression, isTrue);
+      expect('e*x - pi'.isRealFunction, isTrue);
+      expect('e - pi'.isNumericalExpression, isTrue);
+      expect('3*e*x/2'.isRealFunction, isTrue);
+      expect('1+2+3'.isRealFunction, isTrue);
+      expect('cos(pi) - 18 * 6'.isNumericalExpression, isTrue);
+      expect('0'.isNumericalExpression, isTrue);
+      expect('e^(2 + pi)'.isNumericalExpression, isTrue);
+
+      expect('2x+6'.isRealFunction, isFalse);
+      expect(''.isRealFunction, isFalse);
+      expect('x'.isNumericalExpression, isFalse);
+      expect('2+x'.isNumericalExpression, isFalse);
     });
   });
 }